Refactor UserProfileFacilityLists and UserProfileProductionLocations components to use new file structure
This commit updates the import paths for UserProfileFacilityLists and UserProfileProductionLocations components to reflect their new locations. Additionally, it removes the old UserProfileFacilityLists component file and introduces a new styles file for both components, enhancing organization and maintainability.
