+++ src/react/src/__tests__/components/UserProfileProductionLocations.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import renderWithProviders from '../../util/testUtils/renderWithProviders';
-import UserProfileProductionLocations from '../../components/UserProfileProductionLocations';
+import UserProfileProductionLocations from '../../components/UserProfileProductionLocations/UserProfileProductionLocations';
 
 const makeFacility = (id, name, address, countryCode) => ({
     id,
